refactor: 애니 리스트, 로그인, 고객센터, 검색에 헤더 컴포넌트 반영
