+++ src/features/auth/routes/Login/index.tsx
@@ -2,11 +2,12 @@ import { NavArrowLeft } from "iconoir-react";
 
 import Button from "@/components/Button";
 import Head from "@/components/Head";
+import Header from "@/components/Layout/Header";
 
 import useLogin from "../../hooks/useLogin";
 
 import SocialGroup from "./SocialGroup";
-import { Header, HeaderContents, Main, Title, LoginSection } from "./style";
+import { Main, Title, LoginSection } from "./style";
 
 export default function Login() {
   const { handleClickBack, handleSocialLogin } = useLogin();
@@ -18,7 +19,7 @@ export default function Login() {
         description="오덕에 회원가입 없이 빠르게 덕질을 시작해 보세요!"
       />
       <Header>
-        <HeaderContents>
+        <Header.Left>
           <Button
             name="뒤로가기"
             icon={<NavArrowLeft aria-hidden />}
@@ -26,8 +27,10 @@ export default function Login() {
             color="neutral"
             onClick={handleClickBack}
           />
+        </Header.Left>
+        <Header.Center>
           <h1>로그인</h1>
-        </HeaderContents>
+        </Header.Center>
       </Header>
       <Main htmlType={"main"}>
         <Title>
